requirements.txt
certifi==2020.4.5.2
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
pycparser==2.20
python-telegram-bot==12.7
six==1.15.0
toml==0.10.1
tornado==6.0.4
